clean up the input-passing stuff, include controller in passed struct

+++ Packages/org.plunderludics.UnityHawk/Runtime/Input/ConvertInput.cs
@@ -12,7 +12,7 @@ public static class ConvertInput {
         return new Plunderludics.UnityHawk.Shared.InputEvent {
             name = ie.name,
             value = ie.value,
-            controller = ie.controller,
+            controller = (int)ie.controller,
             isAnalog = ie.isAnalog
         };
     }
